% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{en_boggle_dices}
\alias{en_boggle_dices}
\title{English Boggle dices.}
\format{A character vector with sixteen 6-letters strings, each representing
  one dice.}
\usage{
en_boggle_dices
}
\description{
Sixteen dices of the english version of the Boggle game.
Original Boggle dices according to
[https://boardgames.stackexchange.com/questions/48151/boggle-letter-distribution-and-face-orientation-of-each-die](https://boardgames.stackexchange.com/questions/48151/boggle-letter-distribution-and-face-orientation-of-each-die)
}
\keyword{datasets}
