% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_points.R
\name{get_points}
\alias{get_points}
\title{Calculate Boggle words points.}
\usage{
get_points(words)
}
\arguments{
\item{words}{A character vector containing all the found words.}
}
\description{
Given the `words`, calculate the total points obtained.
}
