% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_board.R
\name{new_board}
\alias{new_board}
\title{Creates a random Boggle board.}
\usage{
new_board(m = 4, n = 4, dices = boggler::en_boggle_dices)
}
\arguments{
\item{m}{Number of rows of the Boggle board.}

\item{n}{Number of columns of the Boggle board.}

\item{dices}{Character vector, where each item must be a 6-letters string.
Each item represents a dice.}
}
\description{
Given the `dices`, it creates a new game of Boggle of size `m` * `n`.
}
