% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve}
\alias{solve}
\title{Solve a Boggle board.}
\usage{
solve(boggle_board, dict = spark_intro_dict, word_min_len = 3)
}
\arguments{
\item{boggle_board}{The Boggle board represented as a matrix of letters.}

\item{dict}{A character vector containing all the valid words.}

\item{word_min_len}{A numeric indicating the minimum length, in characters,
of a word to count as a valid word.}
}
\description{
Given a Boggle board, this function will return all words contained in
`dict`, which can be formed from the `boggle_board`.
}
