% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spark_intro_dict}
\alias{spark_intro_dict}
\title{"Mastering Apache Spark with R" Introduction dictionary.}
\format{A character vector with words.}
\usage{
spark_intro_dict
}
\description{
Words from the Introduction chapter from "Mastering Apache Spark with R", to
be used as dictionary.
}
\keyword{datasets}
